requirements.txt
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.28.0
sentencepiece>=0.1.99
safetensors>=0.4.3
huggingface-hub>=0.23.0
opencv-python>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
